refactor: Done Moving the Structure
what the hell
