+++ Plugin.cs
@@ -7,8 +7,11 @@ using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 using BepInEx;
 using HarmonyLib;
+using ResourcefulHands.Assets;
 using ResourcefulHands.Core;
 using ResourcefulHands.Patches;
+using ResourcefulHands.Systems;
+using ResourcefulHands.UI;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 using UnityEngine.UI;
