--- Source/Types/Debug/Logger.cs
@@ -1,15 +1,0 @@
-using System;
-using System.Diagnostics;
-using System.IO;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-
-namespace ResourcefulHands;
-
-/// <summary>
-/// - Use RHLog in place of Debug.Log (for the BepInEx console).
-/// - Use RHLog.Player for player logs when logging to the CommandConsole (for the in-game console).
-/// </summary>
-public static class RHLog
-{
-}
