+++ Systems/CosmeticSystem.cs
@@ -1,4 +1,5 @@
 using ResourcefulHands.Core;
+using ResourcefulHands.UI;
 using UnityEngine;
 
 namespace ResourcefulHands.Systems;
