+++ UI/OF_CosmeticPage.cs
@@ -2,14 +2,16 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
+using ResourcefulHands.Assets;
 using ResourcefulHands.Core;
+using ResourcefulHands.Systems;
 using TMPro;
 using Unity.VisualScripting;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 using UnityEngine.UI;
 
-namespace ResourcefulHands;
+namespace ResourcefulHands.UI;
 
 public class OF_CosmeticPage : MonoBehaviour
 {
